+++ net/include/net/utils.h
@@ -46,6 +46,7 @@ struct VpnEndpoint {
     const char *name;         // endpoint host name (used, for example, for TLS handshake)
     const char *remote_id;    // if not NULL or empty, used for server TLS certificate verification instead of `name`
     AG_ARRAY_OF(uint8_t) additional_data; // additional data about the endpoint
+    bool has_ipv6; // Whether IPv6 traffic can be routed through the endpoint
 };
 
 typedef AG_ARRAY_OF(VpnEndpoint) VpnEndpoints;