+++ standalone/setup_wizard/src/template_settings.rs
@@ -65,6 +65,8 @@ hostname = ""
 {}
 addresses = []
 {}
+has_ipv6 = {}
+{}
 username = ""
 {}
 password = ""
@@ -82,6 +84,8 @@ anti_dpi = false
     Endpoint::doc().to_toml_comment(),
     Endpoint::doc_hostname().to_toml_comment(),
     Endpoint::doc_addresses().to_toml_comment(),
+    Endpoint::doc_has_ipv6().to_toml_comment(),
+    Endpoint::default_has_ipv6(),
     Endpoint::doc_username().to_toml_comment(),
     Endpoint::doc_password().to_toml_comment(),
     Endpoint::doc_skip_verification().to_toml_comment(),